Add files via upload
Simulates a shop as the owner in which you can add, modify, and eliminate products. 
Then it takes you in the journey of a costumer, being able to buy and select a specified quantity, as well as emptying the cart. It provides change at the cash register.
After that it takes you back to the owner's menu and you can close up shop or assist another customer. This continues until the owner closes.